Change thetype from int to float for the variables holding the calculation numbers (num1 & num2). also only print 3 digits after the decimal point.

+++ src/main.c
@@ -6,7 +6,7 @@
 //using argv[2][0] or *argv[2] is the same
 
 int main(int argc, char **argv) {
-	int num1, num2;
+	float num1, num2;
 	if(argc>1 && !strcmp(argv[1], "--help")) {
 		help();
 		return 0;
@@ -15,16 +15,16 @@ int main(int argc, char **argv) {
 		num1=atoi(argv[1]);
 		num2=atoi(argv[3]);
 		if(argv[2][0]=='+') {
-			printf("%d + %d = \e[1m%d\e[0m\n", num1, num2, num1+num2);
+			printf("%.3f + %.3f = \e[1m%.3f\e[0m\n", num1, num2, num1+num2);
 			break;
 		} else if(*argv[2]=='-') {
-			printf("%d - %d = \e[1m%d\e[0m\n", num1, num2, num1-num2);
+			printf("%.3f - %.3f = \e[1m%.3f\e[0m\n", num1, num2, num1-num2);
 			break;
 		} else if(*argv[2]=='x'|| *argv[2]=='*') {
-			printf("%d * %d = \e[1m%d\e[0m\n", num1, num2, num1*num2);
+			printf("%.3f * %.3f = \e[1m%.3f\e[0m\n", num1, num2, num1*num2);
 			break;
 		} else if(*argv[2]=='/') {
-			printf("%d / %d = \e[1m%d\e[0m\n", num1, num2, num1/num2);
+			printf("%.3f / %.3f = \e[1m%.3f\e[0m\n", num1, num2, num1/num2);
 			break;
 		} else {
 			fprintf(stderr, "\e[31m\e[1mInvalid operation!\e[0m\n");
